feat(createKunde): Handle createKunde POST Request
Add repo for kunden, create service interface, and implementing class,
to handle the createKunde request, which is called from the
CustomRestController.

Receiving the name for the new Kunde, will create a new object with the
corresponding name, and the current date. This object is saved to the
database using the repository. Thereafter, we return the object to the
origin of the request.