feat&fix(CustomRestController): Add setKundeBezahlt endpoint
Add new endpoint to change kunde.bezahlt to true for a given id.
Refactor IdClass, to hold an unspecified id instead of a specific id, in
order to reuse the class for multiple endpoints.
Refactor the setBestellungStatusVorbereitet endpoint to fit to the new
IdClass, additionally check the return value for the expected number of
changed rows.

To check if the new endpoint works, add tests/set_kunde_bezahlt.sh
script.